Extract common layout
