+++ webapp/src/components/modes/TimeGame.js
@@ -1,35 +1,25 @@
 "use client"
 
 import { useEffect, useState } from "react"
-import { Toolbar, Typography, Button, Card, CardContent, Grid, Box, Container, LinearProgress, Dialog, DialogActions, DialogContent, DialogTitle } from "@mui/material"
-import HelpOutlineIcon from "@mui/icons-material/HelpOutline"
-import PhoneIcon from "@mui/icons-material/Phone"
-import ChatIcon from "@mui/icons-material/Chat"
-import EmojiEventsIcon from "@mui/icons-material/EmojiEvents"
-import MonetizationOnIcon from "@mui/icons-material/MonetizationOn"
-import InterpreterModeIcon from '@mui/icons-material/InterpreterMode';
-import Chat from "../LLMChat"
+import { Typography, Card, CardContent, Grid, Box, Container, LinearProgress } from "@mui/material"
 import useAxios from "../../hooks/useAxios"
-import GraphComponent from '../lifelines/GraphComponent';
-import CallFriend from "../lifelines/CallFriend";
-import PhoneDialog from "../phone/PhoneDialog";
 import useAuth from "../../hooks/useAuth"
-import { NavLink, useNavigate } from "react-router";
-import { GameContainer, StyledAppBar, LogoButton, ScoreChip, CoinsChip, LifelineButton, OptionButton, ImageContainer, LoadingContainer } from "./BaseStyles";
-import useCoinHandler from "../CoinHandler"
-import useLifeLinesHandler from "../lifelines/LifeLinesHandler"
+import { useNavigate } from "react-router";
+import { OptionButton, ImageContainer, LoadingContainer, HighlightedTopic } from "./BaseStyles";
+import useCoinHandler from "../../handlers/CoinHandler"
+import useLifeLinesHandler from "../../handlers/LifeLinesHandler"
+import { TOPIC_QUESTION_MAP } from "../../utils/topicQuestionMap";
+import BaseGame from "./BaseGameLayout"
 
 function TimeGame() {
   const axios = useAxios();
   const { auth } = useAuth();
   const navigate = useNavigate();
   const [roundData, setRoundData] = useState(null);
-  const [roundPrompt, setRoundPrompt] = useState("");
   const [score, setScore] = useState(0);
   const [selectedAnswer, setSelectedAnswer] = useState(null);
   const [loading, setLoading] = useState(true);
   const [chatKey, setChatKey] = useState(0); // resets the chat component every time it is updated
-  const [questions, setQuestions] = useState([]);
   const [showStatistics, setShowStatistics] = useState(false);
   const [correctAnswers, setCorrectAnswers] = useState(0);
 
@@ -43,9 +33,8 @@ function TimeGame() {
     hiddenOptions, isTrue, setHiddenOptions, setShowGraph, newGame } = useLifeLinesHandler(roundData, spendCoins);
 
   useEffect(() => {
-  
       fetchUserCoins();
-    }, [auth, axios]);
+    }, [auth]);
 
   // Function to load the data for each round.
   const loadRound = async () => {
@@ -91,8 +80,7 @@ function TimeGame() {
   // Check if the game is still loading after modifying the round data
   useEffect(() => {
     if (roundData && roundData.items.length > 0) {
-      let wh = (roundData.topic === "athlete" || roundData.topic === "singer") ? "Who" : "What";
-      setRoundPrompt(`${wh} is this ${roundData.topic}?`);
+      const questionInfo = TOPIC_QUESTION_MAP[roundData.topic] || { wh: "What", name: roundData.topic }; // Value by default if topic is not found
       setLoading(false);
     } else {
       setLoading(true);
@@ -127,7 +115,7 @@ function TimeGame() {
   const endGame = async (questions) => {
     try {
       // Calculate earned coins based on score
-      const earnedCoins = score * 0.3;
+      const earnedCoins = Math.round(score * 0.3);
       await updateUserCoins(earnedCoins);
       await axios.post("/addgame", { username: auth.username, questions });
     } catch (error) {
@@ -140,11 +128,11 @@ function TimeGame() {
     setShowStatistics(false);
     setRoundData(null);
     setScore(0);
-    setQuestions([]);
     newGame();
     setSelectedAnswer(null);
     fetchUserCoins(); // Reset coins to the latest value from the server
     setSpentCoins(0);
+    setCorrectAnswers(0);
 
     setTotalRounds(0)
     setGameEnded(false);
@@ -165,19 +153,6 @@ function TimeGame() {
       setCorrectAnswers(correctAnswers + 1);
       setShowGraph(false);
     }
-
-    let updatedQuestions = [];
-    setQuestions((prev) => {
-      updatedQuestions = [
-        ...prev,
-        {
-          topic: roundData.topic,
-          isCorrect: isCorrect,
-          pointsIncrement: pointsIncrement,
-        },
-      ];
-      return updatedQuestions;
-    });
 
     setTimeout(async () => {
       setTotalRounds(totalRounds+1);
@@ -219,231 +194,104 @@ function TimeGame() {
   }, [gameEnded]);
 
   return (
-    <GameContainer maxWidth="100%" height="100%">
-      {/* Top Bar */}
-      <StyledAppBar position="static">
-        <Toolbar>
-          <LogoButton color="inherit" disableRipple>
-            TRIVIA
-          </LogoButton>
-          <Box sx={{ flexGrow: 1 }} />
-          <Box sx={{ display: "flex", gap: 2 }}>
-            <CoinsChip variant="contained" startIcon={<MonetizationOnIcon />}>
-              {coins} 🪙
-            </CoinsChip>
-            <ScoreChip elevation={0}>
-              <EmojiEventsIcon sx={{ mr: 1 }} />
-              Score: {score}
-            </ScoreChip>
-          </Box>
-        </Toolbar>
-      </StyledAppBar>
-
-      {/* Main Content */}
-      <Grid container spacing={3} sx={{ flex: 1, mt: 2 }}>
-        {/* Left Side (Lifelines) */}
-        <Grid item xs={12} md={3}>
-          <Card elevation={3} >
-            <CardContent>
-              <Typography variant="h4" component="h2" gutterBottom color="primary" sx={{ fontWeight: "bold", textDecoration: "underline" }}>
-                Lifelines
-              </Typography>
-              <LifelineButton
-                variant="contained"
-                startIcon={<HelpOutlineIcon />}
-                onClick={handleFiftyFifty}
-                disabled={isTrue("50") || !canAfford(100)}
-                isUsed={isTrue("50")}
-                colorVariant="blue"
-              >
-                50/50 - 100 🪙 {isTrue("50") && "(Used)"}
-              </LifelineButton>
-
-              <LifelineButton
-                variant="contained"
-                startIcon={<PhoneIcon />}
-                onClick={handleCallFriend}
-                disabled={isTrue("CallFriend")}
-                isUsed={isTrue("CallFriend")}
-                colorVariant="green"
-              >
-                Call a Friend {isTrue("CallFriend") && "(Used)"}
-              </LifelineButton>
-
-              <LifelineButton
-                variant="contained"
-                startIcon={<InterpreterModeIcon />}
-                onClick={handleAudienceCall}
-                disabled={isTrue("AskAudience") || !canAfford(150)}
-                isUsed={isTrue("AskAudience")}
-                colorVariant="red"
-              >
-                Audience Call - 150 🪙 {isTrue("AskAudience") && "(Used)"}
-              </LifelineButton>
-
-              <LifelineButton
-                variant="contained"
-                startIcon={<ChatIcon />}
-                onClick={handleUseChat}
-                disabled={isTrue("UseChat") || !canAfford(200)}
-                isUsed={isTrue("UseChat")}
-                colorVariant="purple"
-              >
-                Use the Chat - 200 🪙 {isTrue("UseChat") && "(Used)"}
-              </LifelineButton>
-
-              <LifelineButton
-                variant="contained"
-                onClick={handlePhoneOut}
-                colorVariant="purple"
-              >
-                Phone Out
-              </LifelineButton>
-
-              {isTrue("CallFriend") && (<CallFriend
-                open={isTrue("CallFriendOpen")}
-                onClose={handleCloseCallFriend}
-                correctAnswer={roundData.itemWithImage.name}
-                possibleAnswers={roundData.items.map(item => item.name)}
-              />)}
-              {isTrue("PhoneOut") && (<PhoneDialog
-                open={isTrue("PhoneOut")}
-                onClose={handlePhoneOutClose}
-                key={chatKey} roundData={roundData}
-              />)}
-
-            </CardContent>{isTrue("ShowGraph") && (
-            <Card elevation={3} sx={{ marginTop: 2, paddingTop: 3 }}>
-              <CardContent>
-                <Typography variant="h4" component="h2" color="primary" sx={{ fontSize: '1.5rem' }}>
-                  The audience says...
+    <BaseGame
+      axios={axios}
+      auth={auth}
+      roundData={roundData}
+      chatKey={chatKey}
+      score={score}
+      correctAnswers={correctAnswers}
+      totalRounds={totalRounds}
+      handleNewGame={handleNewGame}
+      coins={coins}
+      spentCoins={spentCoins}
+      showStatistics={showStatistics}
+      setShowStatistics={setShowStatistics}
+      canAfford={canAfford}
+      handleFiftyFifty={handleFiftyFifty}
+      handleCallFriend={handleCallFriend}
+      handleCloseCallFriend={handleCloseCallFriend}
+      handleAudienceCall={handleAudienceCall}
+      handlePhoneOut={handlePhoneOut}
+      handlePhoneOutClose={handlePhoneOutClose}
+      handleUseChat={handleUseChat}
+      isTrue={isTrue}
+    >
+      {/* Game Area */}
+      <Grid item xs={12} md={6}>
+        <Card elevation={3} sx={{ height: "100%", minHeight: 400 }}>
+          <CardContent>
+            {loading ? (
+              <LoadingContainer>
+                <Typography variant="h6" color="textSecondary" sx={{ mt: 3 }}>
+                  Loading question...
                 </Typography>
-                {roundData && <GraphComponent correctAnswer={roundData.itemWithImage.name}
-                  distractors={roundData.items
-                    .filter(item => item.name !== roundData.itemWithImage.name)
-                    .map(item => item.name)
-                  }
-                />}
-              </CardContent>
-            </Card>
-          )}
-            <CardContent>
-              
-            </CardContent>
-          </Card>
-          
-        </Grid>
-
-        {/* Game Area */}
-        <Grid item xs={12} md={6}>
-          <Card elevation={3} sx={{ height: "100%", minHeight: 400 }}>
-            <CardContent>
-              {loading ? (
-                <LoadingContainer>
-                  <Typography variant="h6" color="textSecondary" sx={{ mt: 3 }}>
-                    Loading question...
-                  </Typography>
-                  <Typography variant="body2" color="textSecondary" sx={{ mt: 1 }}>
-                    Get ready!
-                  </Typography>
-                </LoadingContainer>
-              ) : (
-                roundData && (
-                  <>
-                    <Box sx={{ width: "100%", mb: 3 }}>
-                      <Box sx={{ display: "flex", justifyContent: "space-between", mb: 1 }}>
-                        <Typography variant="h6" fontWeight="bold" color="primary">
-                          Time remaining: {timeLeft}s
-                        </Typography>
-                      </Box>
-                      <LinearProgress
-                        variant="determinate"
-                        value={(timeLeft / TIME) * 100}
-                        sx={{
-                          height: 20,
-                          borderRadius: 5,
-                          backgroundColor: "rgba(0, 0, 0, 0.1)",
-                          "& .MuiLinearProgress-bar": {
-                            borderRadius: 5,
-                            backgroundColor: timeLeft > 30 ? "success.main" : timeLeft > 10 ? "warning.main" : "error.main",
-                          },
-                        }}
-                      />
+                <Typography variant="body2" color="textSecondary" sx={{ mt: 1 }}>
+                  Get ready!
+                </Typography>
+              </LoadingContainer>
+            ) : (
+              roundData && (
+                <>
+                  <Box sx={{ width: "100%", mb: 3 }}>
+                    <Box sx={{ display: "flex", justifyContent: "space-between", mb: 1 }}>
+                      <Typography variant="h6" fontWeight="bold" color="primary">
+                        Time remaining: {timeLeft}s
+                      </Typography>
                     </Box>
-                    <ImageContainer>
-                      <img
-                        src={roundData.itemWithImage.imageUrl || "/placeholder.svg"}
-                        alt={roundData.itemWithImage.imageAltText || "Item image"}
-                      />
-                    </ImageContainer>
-                    <Container sx={{ textAlign: "center", mb: 2 }}>
-                      <Typography data-testid="question-prompt" variant="h5" gutterBottom sx={{ fontWeight: "bold" }}>{roundPrompt}</Typography>
-                    </Container>
-                    <Grid container spacing={2}>
-                      {roundData.items.map((item, index) => (
-                        <Grid item xs={6} key={index}>
-                          <OptionButton
-                            variant="contained"
-                            fullWidth
-                            onClick={() => handleOptionSelect(index)}
-                            disabled={hiddenOptions.includes(index)}
-                            isHidden={hiddenOptions.includes(index)}
-                            hasSelectedAnswer={selectedAnswer !== null}
-                            isSelected={selectedAnswer === index}
-                            isCorrect={correctOption(index)}
-                          >
-                            {item.name}
-                          </OptionButton>
-                        </Grid>
-                      ))}
-                    </Grid>
-                  </>
-                )
-              )}
-            </CardContent>
-          </Card>
-        </Grid>
-
-        {/* Right Side (Chat) */}
-        <Grid item xs={12} md={3}>
-          <Card elevation={3} sx={{ height: "100%" }}>
-            <CardContent>
-              {roundData && <Chat key={chatKey} roundData={roundData} />}
-            </CardContent>
-          </Card>
-        </Grid>
+                    <LinearProgress
+                      variant="determinate"
+                      value={(timeLeft / TIME) * 100}
+                      sx={{
+                        height: 20,
+                        borderRadius: 5,
+                        backgroundColor: "rgba(0, 0, 0, 0.1)",
+                        "& .MuiLinearProgress-bar": {
+                          borderRadius: 5,
+                          backgroundColor: timeLeft > 30 ? "success.main" : timeLeft > 10 ? "warning.main" : "error.main",
+                        },
+                      }}
+                    />
+                  </Box>
+                  <ImageContainer>
+                    <img
+                      src={roundData.itemWithImage.imageUrl || "/placeholder.svg"}
+                      alt={roundData.itemWithImage.imageAltText || "Item image"}
+                    />
+                  </ImageContainer>
+                  <Container sx={{ textAlign: "center", mb: 2 }}>
+                    {roundData && (
+                      <Typography data-testid="question-prompt" variant="h5" gutterBottom sx={{ fontWeight: "bold" }}>
+                        {TOPIC_QUESTION_MAP[roundData.topic]?.wh || "What"} is this <HighlightedTopic>{TOPIC_QUESTION_MAP[roundData.topic]?.name || roundData.topic.toUpperCase()}</HighlightedTopic>?
+                      </Typography>
+                    )}
+                  </Container>
+                  <Grid container spacing={2}>
+                    {roundData.items.map((item, index) => (
+                      <Grid item xs={6} key={index}>
+                        <OptionButton
+                          variant="contained"
+                          fullWidth
+                          onClick={() => handleOptionSelect(index)}
+                          disabled={hiddenOptions.includes(index)}
+                          isHidden={hiddenOptions.includes(index)}
+                          hasSelectedAnswer={selectedAnswer !== null}
+                          isSelected={selectedAnswer === index}
+                          isCorrect={correctOption(index)}
+                        >
+                          {item.name}
+                        </OptionButton>
+                      </Grid>
+                    ))}
+                  </Grid>
+                </>
+              )
+            )}
+          </CardContent>
+        </Card>
       </Grid>
-
-      {/* Game statistics */}
-      <Dialog 
-        open={showStatistics} 
-        onClose={(event, reason) => {
-          // Prevent the user to interact with the rest of the screen when the dialog is shown
-          if (reason !== "backdropClick" && reason !== "escapeKeyDown") {
-            setShowStatistics(false)
-          }
-        }}
-        >
-        <DialogTitle sx={{ fontWeight: "bold", color: "primary.main" }}>Game Over</DialogTitle>
-        <DialogContent>
-          <Typography variant="body1"><b>Final Score:</b> {score}</Typography>
-          <Typography variant="body1"><b>Correct Answers:</b> {correctAnswers} / {totalRounds}</Typography>
-          <Typography variant="body1"><b>Accuracy Rate:</b> {((correctAnswers / totalRounds) * 100).toFixed(2)}%</Typography>
-          <Typography variant="body1" colorVariant= "red"><b>Spent on lifelines:</b> {spentCoins} 🪙</Typography>
-          <Typography variant="body1"><b>Earned from correct answers:</b> {score * 0.3} 🪙</Typography>     
-        </DialogContent>
-        <DialogActions sx={{ justifyContent: "center", pb: 2 }}>
-          <NavLink to="/home">
-            <Button variant="contained" color="secondary">
-              Return Home
-            </Button>
-          </NavLink>
-          <Button variant="contained" color="primary" onClick={handleNewGame}>
-            New Game
-          </Button>
-        </DialogActions>
-      </Dialog>
-    </GameContainer>
+        
+    </BaseGame>
   )
 }
